requirements.txt
Django>=5.0
django-environ==0.11.2
djangorestframework==3.14.0
psycopg2==2.9.9
pytest==7.4.4
pytest-django==4.5.2